src/pmdas/linux,pmcd,podman: fix compilation warnings
From vm01 (i686 Ubuntu 18.04), mostly triggered by building on a
32-bit platform.

+++ src/pmdas/linux_proc/acct.c
@@ -297,7 +297,7 @@ open_and_acct(const char *path, int do_acct)
 
     if (lseek(acct_file.fd, file_stat.st_size, SEEK_SET) < 0) {
 	if (pmDebugOptions.appl3)
-	    pmNotifyErr(LOG_DEBUG, "acct: lseek \"%s\",%zd failed: %s\n", path, file_stat.st_size, pmErrStr_r(-oserror(), errmsg, sizeof(errmsg)));
+	    pmNotifyErr(LOG_DEBUG, "acct: lseek \"%s\",%lld failed: %s\n", path, (long long)file_stat.st_size, pmErrStr_r(-oserror(), errmsg, sizeof(errmsg)));
 	goto err3;
     }
